Service/data-resources/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
psycopg[binary]==3.2.7
minio==7.2.15
python-multipart==0.0.20
tifffile==2025.3.30
imagecodecs==2026.1.1
numpy==2.2.4
pillow==11.1.0
pyproj==3.7.1
pydantic==2.11.4
pydantic-settings==2.8.1
rasterio==1.4.3
rio-cogeo==5.3.6
